+++ spec/project/object/native_target_spec.rb
@@ -199,6 +199,21 @@ module ProjectSpecs
           @target.build_configuration_list.set_setting('TARGET_USER_DEFINED', 'TARGET_USER_DEFINED_VALUE')
           @target.resolved_build_setting('PROJECT_REFERENCE_TARGET', true).should == { 'Release' => 'TARGET_USER_DEFINED_VALUE', 'Debug' => 'TARGET_USER_DEFINED_VALUE' }
         end
+
+        it 'returns the resolved build setting considering environment variables' do
+          ENV['TARGET_REFERENCE_ENVIRONMENT'] = 'ENVIRONMENT_VARIABLE_VALUE'
+          target_xcconfig = @project.new_file(fixture_path('target.xcconfig'))
+          @target.build_configuration_list.build_configurations.each { |build_config| build_config.base_configuration_reference = target_xcconfig }
+          @target.resolved_build_setting('TARGET_REFERENCE_ENVIRONMENT', true).should == { 'Release' => 'ENVIRONMENT_VARIABLE_VALUE', 'Debug' => 'ENVIRONMENT_VARIABLE_VALUE' }
+        end
+
+        it 'returns the resolved build setting considering environment variables and variable substitution' do
+          ENV['DEFINED_IN_ENVIRONMENT'] = 'ENVIRONMENT_VARIABLE_VALUE'
+          target_xcconfig = @project.new_file(fixture_path('target.xcconfig'))
+          @target.build_configuration_list.build_configurations.each { |build_config| build_config.base_configuration_reference = target_xcconfig }
+          expected_value = { 'Release' => 'ENVIRONMENT_VARIABLE_VALUE', 'Debug' => 'ENVIRONMENT_VARIABLE_VALUE' }
+          @target.resolved_build_setting('TARGET_REFERENCE_ENVIRONMENT_SUBSTITUTION', true).should == expected_value
+        end
       end
 
       #----------------------------------------#